+++ include/JSystem/JAudio/JASystem/JASTrackInterrupt.hpp
@@ -1,16 +1,16 @@
 #ifndef JASTRACKINTERRUPT_HPP
 #define JASTRACKINTERRUPT_HPP
 
-#include "types.h"
+#include <dolphin/types.h>
 
 namespace JASystem {
 
 class TIntrMgr {
 public:
 	static void init();
-	static unsigned long request(unsigned long);
-	static void setIntr(unsigned long, void*);
-	static void resetInter(unsigned long);
+	static u32 request(u32);
+	static void setIntr(u32, void*);
+	static void resetInter(u32);
 	static void checkIntr();
 	static void timerProcess();
 };
